move coprocess to its own .h.c
