+++ work_queue/src/Makefile
@@ -13,6 +13,7 @@ SOURCES_LIBRARY = \
 
 SOURCES_WORKER = \
 	work_queue_process.o \
+	work_queue_coprocess.o \
 	work_queue_watcher.o \
 	work_queue_gpus.o
